增加注释,以及若干功能

+++ src/main/java/group/uchain/project/dto/ProjectInfo.java
@@ -66,7 +66,7 @@ public class ProjectInfo implements Serializable{
      */
     @JsonIgnore
     @ApiModelProperty("奖项数量")
-    private Integer number;
+    private Double number;
 
     /**
      * 分数类型
